+++ docs/conf.py
@@ -25,7 +25,7 @@ master_doc = 'index'
 project = u'tailon'
 copyright = u'2013-2016, Georgi Valkov'
 
-release = '1.0.0'
+release = '1.1.0'
 version = release
 
 exclude_patterns = ['_build']